requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==23.1.0
black==23.3.0
cffi==1.15.1
chardet==4.0.0
click==8.1.7
discord.py==1.7.3
fire==0.5.0
idna==3.4
importlib-metadata==6.7.0
isort==5.11.5
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.11.2
platformdirs==3.11.0
pycparser==2.21
PyNaCl==1.4.0
pyserial==3.5
RPi.GPIO==0.7.1
six==1.16.0
termcolor==2.3.0
tomli==2.0.1
typed-ast==1.5.5
typing_extensions==4.7.1
yarl==1.9.2
zipp==3.15.0
